Fix endianness conversion on big-endian platforms
I have to admit that the code does not really make a lot of sense
anymore. But to avoid breaking the API, I have published this fix. A new
major release will be published which has a more sane API.

The problem was that val_to_buf would always convert a value to
little-endian, not to native-endian. As on x86 native=little, this would
not be an issue. But when native=big, we would first convert the values
to little-endian, but assume the value would be big-endian encoded.